+++ packages/node/src/behavior/system/commissioning/CommissioningClient.ts
@@ -80,7 +80,6 @@ import {
     CaseAuthenticatedTag,
     CommissioningFlowType,
     DeviceTypeId,
-    DiscoveryCapabilitiesBitmap,
     DiscoveryCapabilitiesSchema,
     FabricIndex,
     ManualPairingCodeCodec,
@@ -88,7 +87,6 @@ import {
     QrPairingCodeCodec,
     Status,
     StatusResponseError,
-    TypeFromPartialBitSchema,
     VendorId,
 } from "@matter/types";
 import { AdministratorCommissioning } from "@matter/types/clusters/administrator-commissioning";
@@ -1031,12 +1029,6 @@ export namespace CommissioningClient {
          * Defaults to 30 seconds.
          */
         timeout?: Duration;
-
-        /**
-         * Discovery capabilities to use for discovery. These are included in the QR code normally and defined if BLE
-         * is supported for initial commissioning.
-         */
-        discoveryCapabilities?: TypeFromPartialBitSchema<typeof DiscoveryCapabilitiesBitmap>;
 
         /**
          * The initial read/subscription used to populate node data.